Checkpoint H10 wind --> H80 wind results
The Holland10 model was transformed to the Holland80 wind model in order
to see the results of the final run. The differences between the H80 and
H10 is that the exponent used for the gradient velocity.

The results of this are not the same as the regular H80 model, it may be
necessary to go back to the previous H80 model.