Fix unit issues and use rrp as additional wind obs
Using the radius of the last closed isobar (rrp) we now have another
value for the outer wind observation. Estimate the wind speed to be
about 5 m/s there and calculate the exponent of that wind speed at that
radius in order to construct the linear equation that determines your
exponent.