Adjust setrun.py files, run h10 egs, reorg files
The setrun.py files were adjusted to be able to include all storm-types,
but for that particular wind model folder the other storm-types were
commented out.

Each file ran an example for the HOLLAND 2010 Parametrized Wind model.

The files were reorganized and rename to shorter names such as cle, h80,
and h10 following the storm name to indicate the folders wind model.